Refactor client out of Expense application
