+++ docker-compose-expense.yml
@@ -11,7 +11,7 @@ services:
   #   ports:
   #     - 18080:8080
   # expense_proxy_java:
-  #   image: joatmon08/consul-envoy:v1.10.3-v1.18.4
+  #   image: joatmon08/consul-envoy:v1.11.4-v1.20.2
   #   restart: always
   #   environment:
   #     CONSUL_HTTP_ADDR: 10.5.0.2:8500
@@ -22,7 +22,7 @@ services:
   #   command: ["consul", "connect", "envoy","-sidecar-for", "expense-java"]
   #   network_mode: "service:expense-java"
   # expense_agent_java:
-  #   image: consul:1.10.3
+  #   image: consul:1.11.4
   #   restart: always
   #   command: ["consul","agent","-config-file=/config/agent.hcl","-config-dir=/config"]
   #   volumes:
@@ -30,16 +30,14 @@ services:
   #   network_mode: "service:expense-java"
 
   expense-dotnet:
-    image: joatmon08/expense:dotnet
-    environment:
-      ASPNETCORE_ENVIRONMENT: Compose
+    image: joatmon08/expense:dotnet-6.0
     networks:
       vpcbr:
         ipv4_address: 10.5.0.7
     ports:
       - 15001:80
   expense_proxy_dotnet:
-    image: joatmon08/consul-envoy:v1.10.3-v1.18.4
+    image: joatmon08/consul-envoy:v1.11.4-v1.20.2
     restart: always
     environment:
       CONSUL_HTTP_ADDR: 10.5.0.2:8500
